Renamed paths in Profile due to correct TabLinks work
